operational (with test data)
next step is to create a kijiji autos web scraper since kijiji-scraper does not work with kijiji autos
