+++ package.json
@@ -20,7 +20,7 @@
   "homepage": "https://github.com/ttamre/carketplace#readme",
   "dependencies": {
     "body-parser": "^1.20.2",
-    "express": "^4.18.3",
-    "kijiji-scraper": "^6.3.4"
+    "ejs": "^3.1.9",
+    "express": "^4.18.3"
   }
 }
