+++ views/login.ejs
@@ -62,7 +62,7 @@
 				<div class="row">
 					<div class="six columns">
 						<label for="loginEmail">Email/Username</label>
-						<input id="loginEmail"class="u-full-width" type="text">
+						<input name="email" id="loginEmail"class="u-full-width" type="text">
 					</div>
 				</div>
 
@@ -70,14 +70,14 @@
 				<div class="row">
 					<div class="six columns">
 						<label for="loginPassword">Password</label>
-						<input id="loginPassword"class="u-full-width" type="password">
+						<input name="password" id="loginPassword"class="u-full-width" type="password">
 					</div>
 				</div>
 
 				<!-- Buttons -->
-                <input id="loginSubmit" type="button" value="Login">
-				<input id="loginCreate" type="button" value="Create">
-                <input id="loginForgot" type="button" value="Forgot Password?">
+                <input name="submit" id="loginSubmit" type="submit" value="Login">
+				<input name="create" id="loginCreate" type="submit" value="Create">
+                <input name="forgot" id="loginForgot" type="submit" value="Forgot Password?">
 
 			</form>
 		</div>
